+++ lib/src/barcode_scanner.dart
@@ -46,6 +46,8 @@ abstract class BarcodeScannerController {
   /// Whether the device can switch between cameras (for example front vs back)
   Future<bool> get supportsSwitchingCamera async => false;
 
+  bool get isSupported;
+
   /// Whether the device has a torch that is on
   bool get torchState => false;