+++ pkgs/blue_bird_scanner/lib/src/blue_bird_barcode_scanner.dart
@@ -72,6 +72,9 @@ class _BlueBirdController extends BarcodeScannerController {
   final BlueBirdScanner _scanner;
   final bool enabled;
 
+  @override
+  bool get isSupported => enabled;
+
   _BlueBirdController(
     this._scanner, {
     required this.enabled,