+++ pkgs/fast_barcode_scanner/lib/src/fast_barcode_scanner.dart
@@ -70,6 +70,9 @@ class FastBarcodeScanner implements BarcodeScanner {
 
 class FastBarcodeScannerController extends BarcodeScannerController {
   @override
+  bool get isSupported => true;
+
+  @override
   void pause() {
     fbs.CameraController.instance.pauseDetector();
   }