+++ pkgs/honeywell/lib/src/honeywell_barcode_scanner.dart
@@ -73,6 +73,9 @@ class _HoneyWellController extends BarcodeScannerController {
   final HoneywellScanner _scanner;
   final bool enabled;
 
+  @override
+  bool get isSupported => enabled;
+
   _HoneyWellController(
     this._scanner, {
     required this.enabled,