+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.1-dev.x
 
 * Added toggleTorch, torchState, toggleCamera and supportsSwitchingCamera
+* Added method `supportsScanner<T extends BarcodeScanner>()` to check if certain scanner is supported on the device
+* Fix issue where zebra didn't work on the second run
 
 ## 0.0.1-dev.1